[add] sampleのroot componentを作成
